+++ docs/changelog.md
@@ -71,6 +71,11 @@
 
 ## 更新说明详情
 
+1.12.59 fix: fix movie_boxoffice_realtime interface
+
+    1. 修复 movie_boxoffice_realtime 接口
+    2. 修复 bank_fjcf_table_detail 接口
+
 1.12.58 fix: fix stock_ipo_summary_cninfo interface
 
     1. 修复 stock_ipo_summary_cninfo 接口
@@ -3534,6 +3539,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.12.59 fix: fix movie_boxoffice_realtime interface
 
 1.12.58 fix: fix stock_ipo_summary_cninfo interface